ディープリンク機能を削除し、SelectBoundModalのパターンに準拠して再実装
